requirements.txt
google-cloud-language
google-cloud-vision
